# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MobiManager API
  description: Complete API documentation for MobiManager - a mobile shop management SaaS platform
  version: 1.0.0
  contact:
    name: MobiManager Support
servers:
  - url: http://localhost:3000
    description: Local development server
security:
  - cookieAuth: []
paths:
...
  /api/auth/super-admin:
    post:
      summary: Super Admin Login
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  admin: { type: object }
  /api/auth/admin/login:
    post:
      summary: Admin Login
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  admin: { type: object }
  /api/auth/admin/register:
    post:
      summary: Admin Registration
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRegister'
      responses:
        '201':
          description: Registration successful
...
  /api/super-admin/admins/{adminId}:
    get:
      summary: Get Admin Details
      tags: [Super Admin]
      parameters:
        - name: adminId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Admin details
    put:
      summary: Update Admin
      tags: [Super Admin]
      parameters:
        - name: adminId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Admin updated
    delete:
      summary: Delete Admin
      tags: [Super Admin]
      parameters:
        - name: adminId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Admin deleted
...
components:
  schemas:
    AdminRegister:
      type: object
      required: [shopName, ownerName, email, phone, password, confirmPassword, city, state]
      properties:
        shopName: { type: string, minLength: 2, maxLength: 150 }
        ownerName: { type: string, minLength: 2, maxLength: 100 }
        email: { type: string, format: email }
        phone: { type: string, pattern: '^[6-9]\\d{9}$' }
        password: { type: string, minLength: 8 }
        confirmPassword: { type: string }
        city: { type: string, minLength: 2 }
        state: { type: string, minLength: 2 }
        address: { type: string }
        gstNumber: { type: string }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: admin_token